Remove Target protocol
