--- Sources/XcodeProjWrapper/PBXNativeTarget+Target.swift
@@ -1,15 +1,0 @@
-internal import XcodeProj
-
-struct TargetWrapper: Target {
-    let target: PBXNativeTarget
-
-    var name: String {
-        target.name
-    }
-
-    var buildConfigurations: [any BuildConfiguration] {
-        (target.buildConfigurationList?.buildConfigurations ?? []).map { buildConfiguration in
-            BuildConfigurationWrapper(buildConfiguration: buildConfiguration)
-        }
-    }
-}
