+++ Sources/XcodeProjWrapper/XcodeProjWrapper.swift
@@ -6,13 +6,13 @@ public struct XcodeProjWrapper: XcodeProjWrapperProtocol {
     private let path: Path
     private let xcodeProj: XcodeProj
 
-    public var targets: [any Target]
+    public var targets: [Target]
 
     public init(path: String) throws {
         self.path = Path(path)
         self.xcodeProj = try XcodeProj(path: self.path)
         self.targets = xcodeProj.pbxproj.nativeTargets.map { target in
-            TargetWrapper(target: target)
+            Target(target: target)
         }
     }
