+++ Sources/XcodeProjWrapper/XcodeProjWrapperProtocol.swift
@@ -1,7 +1,7 @@
 import Foundation
 
 public protocol XcodeProjWrapperProtocol {
-    var targets: [any Target] { get }
+    var targets: [Target] { get }
 
     mutating func saveChanges() throws
 }
