+++ Sources/XcodeProjWrapperMock/TargetMock.swift
@@ -1,11 +1,3 @@
 import XcodeProjWrapper
 
-public struct TargetMock: Target {
-    public let name: String
-    public let buildConfigurations: [any BuildConfiguration]
-
-    public init(name: String, buildConfigurations: [any BuildConfiguration]) {
-        self.name = name
-        self.buildConfigurations = buildConfigurations
-    }
-}
+public typealias TargetMock = Target
